+++ apps/web/src/App.tsx
@@ -8,6 +8,8 @@ import BlogCreate from './pages/BlogCreate';
 import BlogGroupManage from './pages/BlogGroupManage';
 import BlogList from './pages/blog/BlogList';
 import BlogPost from './pages/blog/BlogPost';
+import ClassicsDetail from './pages/ClassicsDetail';
+import ClassicsList from './pages/ClassicsList';
 import ClimberLab from './pages/ClimberLab';
 import Creator from './pages/Creator';
 import CreatorProfile from './pages/CreatorProfile';
@@ -53,6 +55,10 @@ function RouteTitle() {
       title = '内容详情 | Valley';
     } else if (pathname === '/resources') {
       title = '资源整理 | Valley';
+    } else if (pathname === '/classics') {
+      title = '名著馆 | Valley';
+    } else if (pathname.startsWith('/classic/')) {
+      title = '名著阅读 | Valley';
     } else if (pathname.startsWith('/resource/')) {
       title = '资源详情 | Valley';
     } else if (pathname === '/my-space') {
@@ -129,6 +135,8 @@ function App() {
           <Route path="creator/:code" element={<CreatorProfile />} />
           <Route path="resources" element={<Resources />} />
           <Route path="resource/:id" element={<ResourceDetail />} />
+          <Route path="classics" element={<ClassicsList />} />
+          <Route path="classic/:id" element={<ClassicsDetail />} />
           <Route path="my-space" element={<MySpace />} />
           <Route path="my-space/image-text" element={<ImageTextCreate />} />
           <Route path="my-space/image-text-edit/:id" element={<ImageTextCreate />} />